src/requirements.txt
numpy
sklearn
pyyaml
pandas
seaborn
